+++ app/models/score/criteria/zazo_activity.rb
@@ -1,6 +1,6 @@
 class Score::Criteria::ZazoActivity < Score::Criteria::Base
   def self.ratio
-    4
+    16
   end
 
   def calculate
@@ -10,7 +10,7 @@ class Score::Criteria::ZazoActivity < Score::Criteria::Base
   private
 
   def get_activity
-    api_params = { user: contact.zazo_mkey, attrs: [:active_friends] }
-    DataProviderApi.new(api_params).query(:attributes)['active_friends'].size rescue 0
+    api_params = { user: contact.zazo_mkey, attrs: [:friends] }
+    DataProviderApi.new(api_params).query(:attributes)['friends'].size rescue 0
   end
 end